updated tower creating mode
