+++ src/classes/__init__.py
@@ -2,3 +2,4 @@ from .gold import *
 from .research import *
 from .resource import *
 from .gamedata import *
+from .tower_handler import TowerHandler
