+++ src/classes/tower.py
@@ -1,9 +1,10 @@
 import arcade
+from pathlib import Path
 
 
 class Tower(arcade.Sprite):
     """
-    Interface for Tower classes.
+    Base tower class
 
     Parameters
     ----------
@@ -13,7 +14,7 @@ class Tower(arcade.Sprite):
         end user sees this on their screen.
     level: int
         level of tower, increases damage.
-    attack_cooldown: float
+    attack_cooldown_ms: float
         how much it takes to shoot again after shooting in milliseconds.
     radius: float
         how far can the tower reach, for shooting and recognizing enemies.
@@ -34,7 +35,7 @@ class Tower(arcade.Sprite):
         name: str = "",
         label: str = "",
         level: int = 0,
-        attack_cooldown: float = 0,
+        attack_cooldown_ms: float = 0,
         radius: float = 0,
         radius_splash: float = 0,
         damage_air: float = 0,
@@ -48,7 +49,7 @@ class Tower(arcade.Sprite):
 
         self.level: int = level
 
-        self.attack_cooldown: float = attack_cooldown
+        self.attack_cooldown_ms: float = attack_cooldown_ms
         self.radius: float = radius
         self.radius_splash: float = radius_splash
 
@@ -57,4 +58,6 @@ class Tower(arcade.Sprite):
         self.damage_splash: float = damage_splash
         self.damage_poison: float = damage_poison
 
-        self.texture = arcade.load_texture(r"..\resources\images\anti-air-1.png")
+        self.texture = arcade.load_texture(
+            Path("..") / "resources" / "towers" / self.name
+        )
