+++ src/const/__init__.py
@@ -1,2 +1,3 @@
 from .settings import *
 from .views import *
+from .towers import *
