+++ src/views/map.py
@@ -1,6 +1,8 @@
-from random import random
+from pathlib import Path
+
 import arcade
 import src.const as C
+from src.const import towers
 
 from src.classes import *
 
@@ -21,9 +23,24 @@ class MapView(arcade.View):
         Listen to mouse press event
     """
 
-    def __init__(self):
+    def __init__(self, tiled_name: str, label: str):
         # Inherit parent class
         super().__init__()
+
+        self.tiled_name = tiled_name
+        self.label = label
+        self.enemy_handler = None  # TODO
+        self.tower_handler = TowerHandler()
+        self.gold = None  # TODO
+        self.research = None  # TODO
+
+        self._tile_map = arcade.load_tilemap(
+            Path("..") / "resources" / "maps" / tiled_name
+        )
+        self._scene = arcade.Scene.from_tilemap(self._tile_map)
+        self._paths = self._tile_map.get_tilemap_layer("paths")
+
+        self.tower_handler.build_tower(towers.AntiAirTower)
 
     def on_show(self):
         """Called when switching to this view."""
@@ -31,30 +48,16 @@ class MapView(arcade.View):
 
     def on_draw(self):
         """Draw the map view."""
-
-        # TODO: To remove this. It was for testing only.
-        self.clear()
-        arcade.draw_text(
-            f"Gold {Gold.get()} Research {Research.get()}",
-            arcade.get_window().width / 2,
-            arcade.get_window().height / 2,
-            arcade.color.BLACK,
-            font_size=C.VIEWS.FONT_SIZE
-            * 2
-            * arcade.get_window().height
-            / C.SETTINGS.SCREEN_HEIGHT,
-            anchor_x="center",
-        )
+        self._scene.draw()
+        self.tower_handler.tower_list.draw()
 
     def on_update(self, delta_time: float):
-        # TODO: To remove this. It was for testing only.
-        Gold.increment(1)
-        Research.increment((random() * 10) - (random() * 10))
+        self.tower_handler.tower_list[0].center_x += 1
 
     def on_mouse_press(self, _x, _y, _button, _modifiers):
         """Use a mouse press to advance to the 'game' view."""
         # save_data.GameData.read_data()
-        self.window.show_view(MapView())
+        # self.window.show_view(MapView())
 
     def on_key_press(self, symbol, modifiers):
         """Called whenever a key is pressed."""
